+++ vscode-extension/src/commands/authorize-app.ts
@@ -1,50 +1,52 @@
 import { Uri } from "vscode";
 import { AuthorizationStatusBar } from "../features/AuthorizationStatusBar";
 import { LogLevel, output } from "../features/Output";
-import { AppManifest } from "../lib/AppManifest";
-import { ObjIdConfig } from "../lib/ObjIdConfig";
+import { CONFIG_FILE_NAME, ObjIdConfig } from "../lib/ObjIdConfig";
 import { Backend } from "../lib/Backend";
 import { UI } from "../lib/UI";
-import { deauthorizeApp } from "./deauthorize-app";
 import { Telemetry } from "../lib/Telemetry";
-import { authorization } from '../lib/Authorization';
 import { Git } from "../lib/Git";
+import { ALWorkspace } from "../lib/ALWorkspace";
+import { showDocument } from "../lib/functions";
 
-export const authorizeApp = async (uri: Uri, manifest: AppManifest, repeat: boolean = false) => {
-    output.log(`Authorizing app "${manifest.name}" id ${manifest.id}`, LogLevel.Info);
-
-    Telemetry.instance.log("authorize", manifest.id);
-    const gitUser = await Git.instance.getUserInfo(uri);
-    let response = await Backend.authorizeApp(manifest.id, gitUser.name, gitUser.email, async (response) => {
-        const { error } = response;
-        if (error.statusCode !== 405) return false;
-
-        const result = await UI.authorization.showAlreadyAuthorizedWarning(manifest);
-        if (result === "Yes") {
-            let { authKey } = ObjIdConfig.instance(uri!);
-            if (!authKey) {
-                UI.authorization.showNoKeyError(manifest);
-            }
-            let token = { success: false };
-            await deauthorizeApp(uri, manifest, token);
-            if (token.success) {
-                authorizeApp(uri, manifest, true);
-            }
-        }
-        return true;
-    });
-
-    if (!response || !response.authKey) return;
-
-    ObjIdConfig.instance(uri).authKey = response.authKey;
-    if (repeat) {
-        UI.authorization.showReauthorizedInfo(manifest);
-    } else {
-        UI.authorization.showAuthorizationSuccessfulInfo(manifest);
+export const authorizeApp = async () => {
+    const uris = await ALWorkspace.pickFolder(true, "to authorize") as Uri[];
+    if (!uris) {
+        return;
     }
 
-    await authorization.commitAuthorization(uri, "authorizing");
+    const success = await Git.instance.executeCleanOperation({
+        uris,
+        operation: async (manifest) => {
+            output.log(`Authorizing app "${manifest.name}" id ${manifest.id}`, LogLevel.Info);
 
-    AuthorizationStatusBar.instance.updateStatusBar();
-    authorization.showAuthorizedDoc();
-};
+            Telemetry.instance.log("authorize", manifest.id);
+            const gitUser = await Git.instance.getUserInfo(manifest.ninja.uri);
+            let response = await Backend.authorizeApp(manifest.id, gitUser.name, gitUser.email, async (response) => {
+                const { error } = response;
+                if (error.statusCode !== 405) {
+                    return false;
+                }
+
+                await UI.authorization.showAlreadyAuthorizedError(manifest);
+                return true;
+            });
+
+            if (!response || !response.authKey) {
+                return false;
+            }
+
+            ObjIdConfig.instance(manifest.ninja.uri).authKey = response.authKey;
+            UI.authorization.showAuthorizationSuccessfulInfo(manifest);
+            return true;
+        },
+        getFilesToStage: () => [CONFIG_FILE_NAME],
+        learnMore: () => showDocument("authorization-git"),
+        getCommitMessage: (manifest) => `AL Object ID Ninja app authorization for ${manifest.name}`
+    });
+
+    if (success) {
+        AuthorizationStatusBar.instance.updateStatusBar();
+        showDocument("authorized");
+    }
+}