+++ vscode-extension/src/features/AuthorizationStatusBar.ts
@@ -1,11 +1,12 @@
 import { MarkdownString, StatusBarAlignment, StatusBarItem, window } from "vscode";
 import { ALWorkspace } from "../lib/ALWorkspace";
-import { AppManifest, getManifest } from "../lib/AppManifest";
+import { getManifest } from "../lib/AppManifest";
 import { ObjIdConfig } from "../lib/ObjIdConfig";
 import { EXTENSION_NAME, URLS } from "../lib/constants";
 import { DisposableHolder } from "./DisposableHolder";
 import { Backend } from "../lib/Backend";
 import { AuthorizedAppResponse } from "../lib/BackendTypes";
+import { AppManifest } from "../lib/types";
 
 export class AuthorizationStatusBar extends DisposableHolder {
     private _status: StatusBarItem;