+++ vscode-extension/src/lib/Backend.ts
@@ -8,7 +8,7 @@ import { HttpMethod, Https } from "./Https";
 import { executeWithStopwatchAsync } from "./MeasureTime";
 import { UI } from "./UI";
 import { ConsumptionCache } from '../features/ConsumptionCache';
-import { getLastKnownAppName } from './AppManifest';
+import { getManifestFromAppId } from './AppManifest';
 import { ExplorerTreeDataProvider } from '../features/Explorer/ExplorerTreeDataProvider';
 import { LABELS } from './constants';
 import { env, Uri } from 'vscode';
@@ -122,7 +122,8 @@ async function sendRequest<T>(path: string, method: HttpMethod, data: any = {}, 
             if (appInfo) {
                 const { appId } = data;
                 const { _log, ...consumptions } = appInfo;
-                NotificationsFromLog.instance.updateLog(appId, _log as EventLogEntry[], getLastKnownAppName(appId));
+                const manifest = getManifestFromAppId(appId);
+                NotificationsFromLog.instance.updateLog(appId, _log as EventLogEntry[], manifest.name);
                 if (ConsumptionCache.instance.updateConsumption(appId, consumptions as ConsumptionData)) {
                     ExplorerTreeDataProvider.instance.refresh();
                 }