+++ vscode-extension/src/lib/functions.ts
@@ -1,0 +1,6 @@
+import path = require("path");
+import { commands, Uri } from "vscode";
+
+export function showDocument(document: string) {
+    commands.executeCommand("markdown.showPreview", Uri.file(path.join(__dirname, `../../docs/${document}.md`)));
+}